Don't create external process during test
Calling the external process during the test made it hard to capture
stderr/out when there was a test failure. Instead calling
`sdc_tool.main` with arguments, keeping everything running in the same
python process.

Using `capsys` to capture stderr and stdout for test assertions (see
`test_validate.py`. Assertions of process exiting `0` are removed as
they are handled automatically now by pytest.

Assertions need to be added for the other tests.

Also removed the `commands` folder and created a single `commands`
module.
