+++ src/sdc-tool/api.py
@@ -14,7 +14,6 @@
 
 import re
 import logging
-import json
 
 import requests
 import time
